Move some tests to another file because spec.rs is autogenerated
